+++ mobile/index.ts
@@ -1,2 +1,3 @@
 export * from './Box/Box';
+export * from './Offset/Offset';
 export * from './Text/Text';